Add day 9 - implement a hash map algo
Because the instructions only asked to implement `set()` and `get()`, we
don't really need to store the raw `key` string in the Hashmap. If the
Hashmap had to print the keys of the Hashmap or implemet something to
avoid hash collisions, then I would rethink the `set()` solution to
store a `[key, val]` array at the hash location in the internal
`_values` array.